requirements/base.txt
fastapi==0.42.0
requests==2.22.0
uvicorn==0.11.7
python-dotenv==0.10.3
pytest==5.2.1
pytest-cov==2.8.1
black==19.3b0
fairseq==0.12.2
srsly==2.4.4